recommendationEngine/requirements.txt
pandas
scipy
scikit-learn
numpy
psycopg2-binary
dotenv